pyproject.toml
[project]
name = "medml-pipeline"
version = "0.1.0"
description = "Medical ML pipeline from EDA to production using uv"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "scikit-learn>=1.4.0",
  "xgboost>=2.0.0",
  "matplotlib>=3.8.0",
  "seaborn>=0.13.0",
  "missingno>=0.5.2",
  "scipy>=1.12.0",
  "pyyaml>=6.0.1",
  "joblib>=1.3.2",
  "fastapi>=0.111.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.7.0",
  "httpx>=0.27.0",
  "python-multipart>=0.0.9",
  "mlflow>=2.14.0",
  "dvc>=3.51.0",
]
authors = [
  {name = "SALHI Aymane"}
]

[dependency-groups]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "black>=24.4.0",
  "isort>=5.13.2",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
  "pre-commit>=3.7.1",
  "pandas-stubs>=2.2.2",
  "types-PyYAML>=6.0.12",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
pythonpath = ["."]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_unused_ignores = true
warn_return_any = true
